+++ src/test/compactthin_tests.cpp
@@ -90,12 +90,12 @@ BOOST_AUTO_TEST_CASE(request_block_announcements) {
         auto h = w.requestBlockAnnouncements(node);
 
         // Should send a request for header announcements
-        BOOST_CHECK_EQUAL(1, node.messages.size());
+        BOOST_CHECK_EQUAL(size_t(1), node.messages.size());
         BOOST_CHECK_EQUAL("sendcmpct", node.messages.at(0));
     }
     // hande goes out of scope,
     // should send a request to disable header announcements
-    BOOST_CHECK_EQUAL(2, node.messages.size());
+    BOOST_CHECK_EQUAL(size_t(2), node.messages.size());
     BOOST_CHECK_EQUAL("sendcmpct", node.messages.at(1));
 }
